PAY-65: Add paymentMethod aware interface and trait + Adjust models to use the new trait
